* Remove several warnings due to non explicit type casting in ae_dna.cpp, ae_exp_manager.cpp, ae_mutation.cpp, ae_output_manager.cpp, params.cpp, and aevol_run.cpp. * In ae_dna::inter_GU_ABCDE_to_ACDBE, the last byte of the newly created DNA sequences was not initialized (must be initialized with null caracter).

+++ src/libaevol/ae_mutation.cpp
@@ -700,19 +700,19 @@ void ae_mutation::get_generic_description_string( char * str )
     case INSERT :
     {
       sprintf( str, "%"PRId8" %"PRId32" %"PRId32" %"PRId32" %"PRId32" %"PRId8" %"PRId16" %"PRId16" %"PRId32" %"PRId32"", \
-              (int8_t) _mut_type, _pos[0], (int32_t)-1, (int32_t)-1, (int32_t)-1, (int8_t) -1, (int16_t) -1, (int32_t) -1, _length[0], (int32_t) -1 );
+              (int8_t) _mut_type, _pos[0], (int32_t)-1, (int32_t)-1, (int32_t)-1, (int8_t) -1, (int16_t) -1, (int16_t) -1, _length[0], (int32_t) -1 );
       break;
     }
     case INS_HT :
     {
       sprintf( str, "%"PRId8" %"PRId32" %"PRId32" %"PRId32" %"PRId32" %"PRId8" %"PRId16" %"PRId16" %"PRId32" %"PRId32"", \
-              (int8_t) _mut_type, _pos[0], _pos[1], _pos[2], _pos[3], _sense, _align_score[0], _align_score[1], _length[0], (int32_t) -1 );
+              (int8_t) _mut_type, _pos[0], _pos[1], _pos[2], _pos[3], (int8_t) _sense, _align_score[0], _align_score[1], _length[0], (int32_t) -1 );
       break;
     }
     case REPL_HT :
     {
       sprintf( str, "%"PRId8" %"PRId32" %"PRId32" %"PRId32" %"PRId32" %"PRId8" %"PRId16" %"PRId16" %"PRId32" %"PRId32" ", \
-              (int8_t) _mut_type, _pos[0], _pos[1], _pos[2], _pos[3], _sense, _align_score[0], _align_score[1], _length[1], _length[0] );
+              (int8_t) _mut_type, _pos[0], _pos[1], _pos[2], _pos[3], (int8_t) _sense, _align_score[0], _align_score[1], _length[1], _length[0] );
       break;
     }
     default :
